backend/requirements.txt
# ======================================================
# Backend Dependencies - ScholarLens
# ======================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL driver
asyncpg==0.29.0  # Async PostgreSQL driver (optional)

# API & HTTP
httpx==0.26.0  # For Claude API calls
python-multipart==0.0.6

# Environment & Config
python-dotenv==1.0.0

# AI Integration
anthropic==0.18.1  # Claude API client

# Utilities
pydantic[email]==2.5.3

# Testing (optional - user will implement)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
